+++ akshare/__init__.py
@@ -2324,9 +2324,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.6 fix: fix futures_hog_info interface
 1.9.7 add: add stock_profit_forecast_ths interface
 1.9.8 fix: fix stock_hk_valuation_baidu interface
+1.9.9 add: add macro_shipping_bci interface
 """
 
-__version__ = "1.9.8"
+__version__ = "1.9.9"
 __author__ = "AKFamily"
 
 import sys
@@ -4488,6 +4489,10 @@ from akshare.economic.macro_china import (
     macro_china_real_estate,
     macro_china_qyspjg,
     macro_china_fdi,
+    macro_shipping_bci,
+    macro_shipping_bcti,
+    macro_shipping_bdi,
+    macro_shipping_bpi,
 )
 
 """