+++ docs/changelog.md
@@ -50,6 +50,13 @@
 
 ## 更新说明
 
+1.9.9 add: add macro_shipping_bci interface
+
+    1. 新增 macro_shipping_bci 接口
+    2. 新增 macro_shipping_bdi 接口
+    3. 新增 macro_shipping_bpi 接口
+    4. 新增 macro_shipping_bcti 接口
+
 1.9.8 fix: fix stock_hk_valuation_baidu interface
 
     1. 修复 stock_hk_valuation_baidu 接口
@@ -1981,6 +1988,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.9 add: add macro_shipping_bci interface
 
 1.9.8 fix: fix stock_hk_valuation_baidu interface